+++ providers/capa/standard/capa_suite_test.go
@@ -14,17 +14,11 @@ import (
 	"github.com/giantswarm/clustertest/pkg/logger"
 	"github.com/giantswarm/clustertest/pkg/wait"
 
-	"github.com/giantswarm/cluster-test-suites/common"
+	"github.com/giantswarm/cluster-test-suites/internal/state"
 	"github.com/giantswarm/cluster-test-suites/providers/capa"
 )
 
 const KubeContext = "capa"
-
-var (
-	ctx       context.Context
-	framework *clustertest.Framework
-	cluster   *application.Cluster
-)
 
 func TestCAPAStandard(t *testing.T) {
 	RegisterFailHandler(Fail)
@@ -32,21 +26,20 @@ func TestCAPAStandard(t *testing.T) {
 }
 
 var _ = BeforeSuite(func() {
-	ctx = context.Background()
 	logger.LogWriter = GinkgoWriter
 
-	var err error
-	framework, err = clustertest.New(KubeContext)
+	state.SetContext(context.Background())
+
+	framework, err := clustertest.New(KubeContext)
 	Expect(err).NotTo(HaveOccurred())
+	state.SetFramework(framework)
 
-	cluster = setUpWorkloadCluster()
-
-	common.Framework = framework
-	common.Cluster = cluster
+	cluster := setUpWorkloadCluster()
+	state.SetCluster(cluster)
 })
 
 func setUpWorkloadCluster() *application.Cluster {
-	cluster, err := framework.LoadCluster()
+	cluster, err := state.GetFramework().LoadCluster()
 	Expect(err).NotTo(HaveOccurred())
 	if cluster != nil {
 		logger.Log("Using existing cluster %s/%s", cluster.Name, cluster.Namespace)
@@ -57,23 +50,24 @@ func setUpWorkloadCluster() *application.Cluster {
 }
 
 func createCluster() *application.Cluster {
-	cluster = capa.NewClusterApp("", "", "./test_data/cluster_values.yaml", "./test_data/default-apps_values.yaml")
+	cluster := capa.NewClusterApp("", "", "./test_data/cluster_values.yaml", "./test_data/default-apps_values.yaml")
 	logger.Log("Workload cluster name: %s", cluster.Name)
+	state.SetCluster(cluster)
 
-	applyCtx, cancelApplyCtx := context.WithTimeout(ctx, 20*time.Minute)
+	applyCtx, cancelApplyCtx := context.WithTimeout(state.GetContext(), 20*time.Minute)
 	defer cancelApplyCtx()
 
-	client, err := framework.ApplyCluster(applyCtx, cluster)
+	client, err := state.GetFramework().ApplyCluster(applyCtx, state.GetCluster())
 	Expect(err).NotTo(HaveOccurred())
 
 	Eventually(
-		wait.AreNumNodesReady(ctx, client, 1, &cr.MatchingLabels{"node-role.kubernetes.io/control-plane": ""}),
+		wait.AreNumNodesReady(state.GetContext(), client, 1, &cr.MatchingLabels{"node-role.kubernetes.io/control-plane": ""}),
 		20*time.Minute, 15*time.Second,
 	).Should(BeTrue())
 
 	DeferCleanup(func() {
-		Expect(framework.DeleteCluster(ctx, cluster)).To(Succeed())
+		Expect(state.GetFramework().DeleteCluster(state.GetContext(), state.GetCluster())).To(Succeed())
 	})
 
-	return cluster
+	return state.GetCluster()
 }